requirements.txt
cassandra-driver==3.28.0
mysql-connector-python==8.1.0
pandas~=2.0.3